+++ frontend/app/src/components/history/MovementLinks.vue
@@ -4,26 +4,22 @@
       <span class="mr-1 font-weight-medium">
         {{ $t('movement_links.address') }}
       </span>
-      <hash-link :text="item.address" :base-url="addressBaseUrl" full-address />
+      <hash-link :text="item.address" :chain="chain" full-address />
     </span>
     <span v-if="item.transactionId" class="d-flex flex-row mt-1">
       <span class="mr-1 font-weight-medium">
         {{ $t('movement_links.transaction') }}
       </span>
-      <hash-link
-        :text="transactionId"
-        :base-url="transactionBaseUrl"
-        full-address
-      />
+      <hash-link :text="transactionId" :chain="chain" tx full-address />
     </span>
   </span>
 </template>
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import { mapGetters } from 'vuex';
-import { explorerUrls } from '@/components/helper/asset-urls';
 import HashLink from '@/components/helper/HashLink.vue';
 import { AssetMovement } from '@/services/history/types';
+import { ETH } from '@/typing/types';
 
 @Component({
   components: { HashLink },
@@ -31,14 +27,13 @@ import { AssetMovement } from '@/services/history/types';
 })
 export default class MovementLinks extends Vue {
   isEthereumToken!: (asset: string) => boolean;
-  readonly urls = explorerUrls;
 
   @Prop({ required: true })
   item!: AssetMovement;
 
   get transactionId(): string {
     const { transactionId } = this.item;
-    if (!this.isEthOrToken) {
+    if (this.chain !== ETH) {
       return transactionId;
     }
     return transactionId.startsWith('0x')
@@ -46,38 +41,11 @@ export default class MovementLinks extends Vue {
       : `0x${transactionId}`;
   }
 
-  get isEtc(): boolean {
-    return this.item.asset === 'ETC';
-  }
-
-  get isBtc(): boolean {
-    return this.item.asset === 'BTC';
-  }
-
-  get isEthOrToken(): boolean {
-    return this.item.asset === 'ETH' || this.isEthereumToken(this.item.asset);
-  }
-
-  get addressBaseUrl(): string | null {
-    if (this.isBtc) {
-      return this.urls.BTC.address;
-    } else if (this.isEthOrToken) {
-      return this.urls.ETH.address;
-    } else if (this.isEtc) {
-      return this.urls.ETC.address;
+  get chain(): string {
+    if (this.isEthereumToken(this.item.asset) || this.item.asset === ETH) {
+      return ETH;
     }
-    return null;
-  }
-
-  get transactionBaseUrl(): string | null {
-    if (this.isBtc) {
-      return this.urls.BTC.transaction;
-    } else if (this.isEthOrToken) {
-      return this.urls.ETH.transaction;
-    } else if (this.isEtc) {
-      return this.urls.ETC.transaction;
-    }
-    return null;
+    return this.item.asset;
   }
 }
 </script>