+++ frontend/app/tests/unit/store/settings/mutations.spec.ts
@@ -10,7 +10,8 @@ import {
   TIMEFRAME_SETTING,
   TIMEFRAME_TWO_WEEKS,
   TIMEFRAME_YEAR,
-  REFRESH_PERIOD
+  REFRESH_PERIOD,
+  EXPLORERS
 } from '@/store/settings/consts';
 import { SettingsState } from '@/store/settings/types';
 import store from '@/store/store';
@@ -30,7 +31,12 @@ describe('settings:mutations', () => {
       [CURRENCY_LOCATION]: CURRENCY_BEFORE,
       [THOUSAND_SEPARATOR]: '|',
       [DECIMAL_SEPARATOR]: '-',
-      [REFRESH_PERIOD]: 120
+      [REFRESH_PERIOD]: 120,
+      [EXPLORERS]: {
+        ETH: {
+          transaction: 'explore/tx'
+        }
+      }
     };
     store.commit('settings/restore', state);
     const settings = store.state.settings!;
@@ -46,5 +52,10 @@ describe('settings:mutations', () => {
     expect(settings[DECIMAL_SEPARATOR]).toBe('-');
     expect(settings[CURRENCY_LOCATION]).toBe(CURRENCY_BEFORE);
     expect(settings[REFRESH_PERIOD]).toBe(120);
+    expect(settings[EXPLORERS]).toStrictEqual({
+      ETH: {
+        transaction: 'explore/tx'
+      }
+    });
   });
 });