+++ frontend/app/tests/unit/store/settings/utils.spec.ts
@@ -11,7 +11,8 @@ import {
   THOUSAND_SEPARATOR,
   DECIMAL_SEPARATOR,
   CURRENCY_LOCATION,
-  REFRESH_PERIOD
+  REFRESH_PERIOD,
+  EXPLORERS
 } from '@/store/settings/consts';
 import { loadFrontendSettings } from '@/store/settings/utils';
 
@@ -62,7 +63,8 @@ describe('settings:utils', () => {
         [THOUSAND_SEPARATOR]: Defaults.DEFAULT_THOUSAND_SEPARATOR,
         [DECIMAL_SEPARATOR]: Defaults.DEFAULT_DECIMAL_SEPARATOR,
         [CURRENCY_LOCATION]: Defaults.DEFAULT_CURRENCY_LOCATION,
-        [REFRESH_PERIOD]: -1
+        [REFRESH_PERIOD]: -1,
+        [EXPLORERS]: {}
       },
       { root: true }
     );